@@ -168,21 +169,29 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^28.1.3",
     "kubo": "^0.24.0",
+    "mini-css-extract-plugin": "^2.8.0",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
     "os-browserify": "^0.3.0",
     "patch-package": "^6.4.7",
     "path-browserify": "^1.0.1",
     "playwright-chromium": "^1.27.1",
+    "postcss": "^8.4.35",
     "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-loader": "^8.1.0",
     "run-script-os": "^1.1.6",
     "semantic-release": "^19.0.3",
     "stream-browserify": "^3.0.0",
-    "typescript": "^4.1.3",
+    "style-loader": "^3.3.4",
+    "ts-loader": "^9.5.1",
+    "typescript": "^4.9.5",
     "uint8arrays": "^4.0.9",
     "wait-on": "^6.0.1",
-    "webpack": "^5.73.0",
-    "webpack-bundle-analyzer": "^3.8.0"
+    "webpack": "^5.90.1",
+    "webpack-bundle-analyzer": "^3.8.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "workbox-webpack-plugin": "^7.0.0"
   },
   "homepage": "./",
   "bundlesize": [
